Otprilike tacka od koje bih mogao da se vratim posle, u slucaju da ne stignem da zavrsim projekat sada u ovom roku

+++ backend/routes/posts.js
@@ -4,6 +4,8 @@ const multer = require("multer");
 
 
 const Post = require('../models/post');
+const checkAuth = require("../middleware/check-auth");
+
 
 const router = express.Router();
 
@@ -32,7 +34,7 @@ const storage = multer.diskStorage( {
 });
 
 
-router.put("/:id",multer({storage: storage}).single("image"),(req,res,next) => {
+router.put("/:id",checkAuth,multer({storage: storage}).single("image"),(req,res,next) => {
   let imagePath = req.body.imagePath;
   if(req.file) {
     const url = req.protocol + "://" +req.get("host");
@@ -42,23 +44,30 @@ router.put("/:id",multer({storage: storage}).single("image"),(req,res,next) => {
     _id : req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Update succesful!'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified >0) {
+      res.status(200).json({message: 'Update succesful!'});
+    }
+    else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   })
 })
 
 
 
 
-router.post("", multer({storage: storage}).single("image"),(req,res,next) => {
+router.post("",checkAuth, multer({storage: storage}).single("image"),(req,res,next) => {
   const url = req.protocol + "://" +req.get("host");
   const post = new Post( {
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -80,7 +89,6 @@ router.get('',(req,res,next) => {
   if(pageSize && currentPage) {
     postQuery.skip(pageSize *(currentPage - 1)).limit(pageSize);
   }
-  console.log(req.query);
   postQuery.find()
     .then(documents => {
       fetchedPosts = documents;
@@ -105,10 +113,15 @@ router.get("/:id",(req,res,next) => {
   })
 });
 
-router.delete("/:id", (req,res,next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+router.delete("/:id",checkAuth, (req,res,next) => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Post deleted!'});
+    if(result.n >0) {
+      res.status(200).json({message: 'Deletion  succesful!'});
+    }
+    else {
+      res.status(401).json({message: "Not authorized!"});
+    }
   })
 
 }); //extrakovace ekspres
